requirements.txt
Flask==3.0.3
openai==1.34.0
pymongo==4.5.0
httpx==0.27.0
langgraph==0.0.64
python-dotenv==1.0.1
langchain==0.2.2
langchain-community==0.2.3
langchain-core==0.2.4
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langchainhub==0.1.20
langcodes==3.3.0
langgraph==0.0.64
langsmith==0.1.74
lazy_loader==0.3
libclang==16.0.0
azure-cosmos==4.3.0
websockets
psutil==5.9.5
bcrypt
fuzzywuzzy
python-Levenshtein
Flask-Mail
PyMuPDF
Flask-CORS
requests